Add bacteria examples and contour drawing
Added a bacteria exampels that shows a single bacteria
its samplinga and its image with and without boundary.

Added contour drawing properties to microscope model.
Uses the same principles as the test script.
Had to add spline and boundary properties to Bacteria.
Fixed a few bugs regarding sampling and how the contours
are drawn.