Change sampling of bacteria
Instead of using rejection sampling within a uniform cube
sample uniformly a torus and a sphere sepereatley both
by rejection sampling. The sphere is sampled by rejection
sampling as it is efficient. The torus is sampled with with
-theta_max to theta_max uniformly phi from 0 to 2pi and
r is R*sqrt(uniform(0,1)). Also we have to discard sample samples
to ensure that the spread is unifrom and that regions of small
curvature have the same density of samples: Sampling from a
Manifold (Holmes et al)